Removed duplicate information
This "Note: If you are on Enterprise managed network, and having issues downloading Mixed Reality Portal, it could be because your organization is using Windows Server Update Services (WSUS) or has other policies that may block the download. You will need to contact your organizations IT department or system administrator to enable Windows Mixed Reality" was repeated already in the third question.